+++ lib/radix_encoding.rb
@@ -4,6 +4,12 @@ require "radix_encoding/encoding"
 require "radix_encoding/version"
 
 module RadixEncoding
+  # Standard Base16 encoding as defined in RFC 4648.
+  BASE16 = Encoding.new(
+    alphabet: "0123456789ABCDEF",
+    radix: 16,
+  )
+
   # Standard Base32 encoding as defined in RFC 4648.
   BASE32 = Encoding.new(
     alphabet: "ABCDEFGHIJKLMNOPQRSTUVWXYZ234567",