+++ spec/spec_helper.rb
@@ -1,7 +1,6 @@
 # frozen_string_literal: true
 
 require "bundler/setup"
-require "radix_encoding"
 
 RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure